Add links to original components and DDF implementation
